Hello World
The hello world of this boilereplate, future updates will include further configuration in threeconfig, typescript, vr configuration built on top of the threeconfig, and further abstraction of important 3d operations such as importing and exporting objects, a easy to use GUI interface for text and ui's inside the threejs canvas, and more. Stay tuned.